Not-really-needed commit to retry CI.
The first build of this package didn't trigger AppVeyor, and so we don't have
a Windows version of this package. I don't know if a PR will fix that, but
it's worth a try. This commit updates `bld.sh` and `build.sh` based on my
subsequent work on getting X.org to build on Windows. The changes aren't
needed for this simple package, but as long as I'm looking for an excuse to
make a commit, I might as well incorporate them.